+++ billy/customer/models.py
@@ -6,6 +6,7 @@ from pytz import UTC
 from datetime import datetime
 from billy.errors import AlreadyExistsError, NotFoundError, LimitReachedError
 from billy.coupons.models import Coupon
+from billy.plans.models import Plan
 from billy.invoices.models import PlanInvoice, PayoutInvoice
 from decimal import Decimal
 from dateutil.relativedelta import relativedelta
@@ -146,6 +147,122 @@ class Customer(Base):
         if not exists:
             raise NotFoundError('Customer not found. Try different id')
 
+    @classmethod
+    def change_customer_plan(cls, customer_id, marketplace, plan_id,
+                    quantity = 1, charge_at_period_end=False):
+        """
+        Changes a customer's plan
+        :param customer_id: A unique id/uri for the customer
+        :param marketplace: a group/marketplace id/uri the user should be placed
+        :param plan_id: The plan id to assocaite customer with
+        :raise:
+        """
+        customer_obj = cls.retrieve_customer(customer_id, marketplace)
+        plan_obj = Plan.retrieve_plan(plan_id, marketplace, active_only=True)
+        current_coupon = customer_obj.coupon
+        start_date = datetime.now(UTC)
+        due_on = datetime.now(UTC)
+        coupon_use_count = customer_obj.coupon_use.get(current_coupon
+                                                       .coupon_id, 0)
+        use_coupon = True if current_coupon.repeating == -1 or \
+                             coupon_use_count \
+                             <= current_coupon.repeating else False
+        can_trial = True if customer_obj.plan_use.get(plan_obj.plan_id,
+                                                      0) == 0 else False
+        end_date = start_date + plan_obj.to_relativedelta(plan_obj.plan_interval)
+        trial_interval = plan_obj.to_relativedelta(plan_obj.trial_interval)
+        if can_trial:
+            end_date += trial_interval
+            due_on += trial_interval
+        if charge_at_period_end:
+            due_on = end_date
+        if quantity < 1:
+            raise ValueError("Quanity must be greater than 1")
+        amount_base = plan_obj.price_cents * Decimal(quantity)
+        amount_after_coupon = amount_base
+        amount_paid = 0
+        coupon_id = current_coupon.coupon_id if current_coupon else None
+        if use_coupon and current_coupon:
+            dollars_off = current_coupon.price_off_cents
+            percent_off = current_coupon.percent_off_int
+            amount_after_coupon -= dollars_off #BOTH CENTS, safe
+            amount_after_coupon -= int(
+                amount_after_coupon * Decimal(percent_off) / Decimal(100))
+        balance = amount_after_coupon
+        PlanInvoice.create_invoice(customer_id, marketplace, plan_id, coupon_id,
+                                   start_date, end_date, due_on, amount_base,
+                                   amount_after_coupon, amount_paid, balance,
+                                   quantity, charge_at_period_end)
+        customer_obj.coupon_use[coupon_id] += 1
+        cls.prorate_last_invoice(customer_obj.customer_id, marketplace, plan_id)
+        cls.session.commit()
+        return customer_obj
+
+
+    def cancel_plan(self, plan_id, cancel_at_period_end=False):
+        """
+        Cancels the customers subscription. You can either do it immediately
+        or at the end of the period.
+         in (matches balanced payments marketplaces)
+        :param at_period_end: Whether to cancel now or wait till the user
+        :returns: New customer object.
+        """
+        if cancel_at_period_end:
+            result = PlanInvoice.retrieve_invoice(self.customer_id,
+                                                self.marketplace,
+                                                plan_id, active_only=True)
+            result.active = False
+            self.session.commit()
+        else:
+            self.prorate_last_invoice(self.customer_id, self.marketplace,
+                                      plan_id)
+        return True
+
+    @classmethod
+    def cancel_customer_plan(cls, customer_id, marketplace,
+                             cancel_at_period_end=True):
+        """
+        Cancels a customers subscription. You can either do it immediately or
+        at the end of the period.
+        :param customer_id: A unique id/uri for the customer
+        :param marketplace: a group/marketplace id/uri the user should be placed
+        in (matches balanced payments marketplaces)
+        :param cancel_at_period_end: Whether to cancel now or wait till the user
+        :returns: New customer object.
+        """
+        cust_obj = cls.retrieve_customer(customer_id, marketplace)
+        return cust_obj.cancel_plan(cancel_at_period_end, cancel_at_period_end)
+
+    @classmethod
+    def prorate_last_invoice(cls, customer_id, marketplace, plan_id):
+        """
+        Prorates the last invoice when changing a users plan. Only use when
+        changing a users plan.
+        :param customer_id: A unique id/uri for the customer
+        :param marketplace: a group/marketplace id/uri the user should be placed
+        in (matches balanced payments marketplaces)
+        """
+        right_now = datetime.now(UTC)
+        last_invoice = PlanInvoice.retrieve_invoice(customer_id, marketplace,
+                                              plan_id, active_only=True)
+        if last_invoice:
+            time_total = Decimal(
+                (last_invoice.end_dt - last_invoice.due_dt).total_seconds())
+            time_used = Decimal((last_invoice.start_dt - right_now).total_seconds())
+            percent_used = time_used / time_total
+            new_base_amount = last_invoice.amount_base_cents * percent_used
+            new_coupon_amount = last_invoice.amount_after_coupon_cents * \
+                                percent_used
+            new_balance = last_invoice.amount_after_coupon_cents - last_invoice \
+                .amount_paid_cents
+            last_invoice.amount_base_cents = new_base_amount
+            last_invoice.amount_after_coupon_cents = new_coupon_amount
+            last_invoice.remaining_balance_cents = new_balance
+            last_invoice.end_dt = right_now - relativedelta(
+                seconds=30) #Extra safety for find query
+            last_invoice.active = False
+        cls.session.commit()
+
 
     #Todo.. non static
     @classmethod
@@ -216,8 +333,3 @@ class Customer(Base):
         customer.current_payout = None
         Customer.session.commit()
         return customer
-
-
-    def change_plan_qty(self):
-        #Todo
-        pass
