--- billy/invoices/utils.py
@@ -1,18 +1,0 @@
-from billy.invoices.models import PlanInvoice
-
-#TODO!
-def create_invoice():
-    """
-    Shouldn't create invoice directly. Create them through changing user's plan
-    from the customer utils
-    """
-    raise NotImplementedError("DO not create invoices directly.")
-
-def retrieve_invoice():
-    pass
-
-def list_invoices():
-    pass
-
-def process_customer():
-    pass
